+++ irs-api/src/test/java/org/eclipse/tractusx/irs/controllers/BatchControllerTest.java
@@ -34,6 +34,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import java.util.UUID;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.eclipse.tractusx.irs.common.auth.IrsRoles;
 import org.eclipse.tractusx.irs.component.BatchOrderResponse;
 import org.eclipse.tractusx.irs.component.BatchResponse;
 import org.eclipse.tractusx.irs.component.RegisterBatchOrder;
@@ -42,7 +43,6 @@ import org.eclipse.tractusx.irs.services.AuthorizationService;
 import org.eclipse.tractusx.irs.services.CreationBatchService;
 import org.eclipse.tractusx.irs.services.QueryBatchService;
 import org.eclipse.tractusx.irs.services.timeouts.CancelBatchProcessingService;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -80,7 +80,7 @@ class BatchControllerTest {
     }
 
     @Test
-    @WithMockUser(authorities = "view_irs")
+    @WithMockUser(authorities = IrsRoles.VIEW_IRS)
     void shouldReturnBadRequestWhenGlobalAssetIdWithWrongFormat() throws Exception {
         when(authorizationService.verifyBpn()).thenReturn(Boolean.TRUE);
 
@@ -91,7 +91,7 @@ class BatchControllerTest {
     }
 
     @Test
-    @WithMockUser(authorities = "view_irs")
+    @WithMockUser(authorities = IrsRoles.VIEW_IRS)
     void shouldReturnBadRequestWhenBatchSizeNotMod10Compliant() throws Exception {
         when(authorizationService.verifyBpn()).thenReturn(Boolean.TRUE);
 
@@ -103,7 +103,7 @@ class BatchControllerTest {
     }
 
     @Test
-    @WithMockUser(authorities = "view_irs")
+    @WithMockUser(authorities = IrsRoles.VIEW_IRS)
     void shouldRegisterBatchOrder() throws Exception {
         when(authorizationService.verifyBpn()).thenReturn(Boolean.TRUE);
 
@@ -114,7 +114,7 @@ class BatchControllerTest {
     }
 
     @Test
-    @WithMockUser(authorities = "view_irs")
+    @WithMockUser(authorities = IrsRoles.VIEW_IRS)
     void shouldReturnBatchOrder() throws Exception {
         when(authorizationService.verifyBpn()).thenReturn(Boolean.TRUE);
 
@@ -127,7 +127,7 @@ class BatchControllerTest {
     }
 
     @Test
-    @WithMockUser(authorities = "view_irs")
+    @WithMockUser(authorities = IrsRoles.VIEW_IRS)
     void shouldReturnBatch() throws Exception {
         final UUID orderId = UUID.randomUUID();
         final UUID batchId = UUID.randomUUID();
@@ -142,7 +142,7 @@ class BatchControllerTest {
     }
 
     @Test
-    @WithMockUser(authorities = "view_irs")
+    @WithMockUser(authorities = IrsRoles.VIEW_IRS)
     void shouldCancelBatchOrder() throws Exception {
         when(authorizationService.verifyBpn()).thenReturn(Boolean.TRUE);