+++ irs-registry-client/src/main/java/org/eclipse/tractusx/irs/registryclient/DefaultConfiguration.java
@@ -114,8 +114,10 @@ public class DefaultConfiguration {
     @Bean
     @ConditionalOnProperty(prefix = CONFIG_PREFIX, name = CONFIG_FIELD_TYPE, havingValue = CONFIG_VALUE_DECENTRAL)
     public DecentralDigitalTwinRegistryClient decentralDigitalTwinRegistryClient(
-            @Qualifier(EDC_REST_TEMPLATE) final RestTemplate edcRestTemplate) {
-        return new DecentralDigitalTwinRegistryClient(edcRestTemplate);
+            @Qualifier(EDC_REST_TEMPLATE) final RestTemplate edcRestTemplate,
+            @Value("${digitalTwinRegistryClient.shellDescriptorTemplate:}") final String shellDescriptorTemplate,
+            @Value("${digitalTwinRegistryClient.lookupShellsTemplate:}") final String lookupShellsTemplate) {
+        return new DecentralDigitalTwinRegistryClient(edcRestTemplate, shellDescriptorTemplate, lookupShellsTemplate);
     }
 
     @Bean