+++ irs-registry-client/src/test/java/org/eclipse/tractusx/irs/registryclient/decentral/DecentralDigitalTwinRegistryClientTest.java
@@ -1,7 +1,6 @@
 package org.eclipse.tractusx.irs.registryclient.decentral;
 
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -9,24 +8,32 @@ import static org.mockito.Mockito.when;
 
 import java.net.URI;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Optional;
 
 import org.eclipse.edc.spi.types.domain.edr.EndpointDataReference;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestTemplate;
 
 class DecentralDigitalTwinRegistryClientTest {
 
+    public static final String SHELL_DESCRIPTORS = "/shell-descriptors/{aasIdentifier}";
+    public static final String LOOKUP_SHELLS = "/lookup/shells";
     RestTemplate restTemplate = mock(RestTemplate.class);
+    private DecentralDigitalTwinRegistryClient client;
+
+    @BeforeEach
+    void setUp() {
+        client = new DecentralDigitalTwinRegistryClient(restTemplate, SHELL_DESCRIPTORS, LOOKUP_SHELLS);
+    }
 
     @Test
     void shouldCallForAssetAdministrationShellDescriptor() {
         // given
-        DecentralDigitalTwinRegistryClient client = new DecentralDigitalTwinRegistryClient(restTemplate);
         EndpointDataReference endpointDataReference = EndpointDataReference.Builder.newInstance()
                                                                                    .endpoint("url.to.host")
                                                                                    .build();
@@ -45,18 +52,17 @@ class DecentralDigitalTwinRegistryClientTest {
     @Test
     void shouldCallForAllAssetAdministrationShellIdsByAssetLink() {
         // given
-        DecentralDigitalTwinRegistryClient client = new DecentralDigitalTwinRegistryClient(restTemplate);
         EndpointDataReference endpointDataReference = EndpointDataReference.Builder.newInstance()
                                                                                    .endpoint("url.to.host")
                                                                                    .build();
-        when(restTemplate.exchange(anyString(), eq(HttpMethod.POST), any(),
-                any(ParameterizedTypeReference.class))).thenReturn(ResponseEntity.of(Optional.of(new ArrayList<>())));
+        when(restTemplate.exchange(any(), eq(HttpMethod.GET), any(), eq(LookupShellsResponse.class))).thenReturn(
+                ResponseEntity.of(Optional.of(LookupShellsResponse.builder().result(Collections.emptyList()).build())));
 
         // when
         client.getAllAssetAdministrationShellIdsByAssetLink(endpointDataReference, new ArrayList<>());
 
         // then
-        verify(restTemplate).exchange(anyString(), eq(HttpMethod.POST), any(), any(ParameterizedTypeReference.class));
+        verify(restTemplate).exchange(any(), eq(HttpMethod.GET), any(), eq(LookupShellsResponse.class));
     }
 
 }