+++ local/testing/testdata/transform-and-upload.py
@@ -3,11 +3,10 @@
 import argparse
 import json
 import math
+import requests
 import time
 import uuid
 from copy import copy
-
-import requests
 from requests.adapters import HTTPAdapter, Retry
 
 
@@ -150,7 +149,8 @@ def create_aas_shell_3_0(global_asset_id_, id_short_, identification_, specific_
     })
 
 
-def create_submodel_descriptor_3_0(id_short_, identification_, semantic_id_, dataplane_asset_address_, asset_id_, endpoint_):
+def create_submodel_descriptor_3_0(id_short_, identification_, semantic_id_, dataplane_asset_address_, asset_id_,
+                                   endpoint_):
     return json.dumps(
         {
             "description": [],
@@ -174,7 +174,14 @@ def create_submodel_descriptor_3_0(id_short_, identification_, semantic_id_, dat
                         "endpointProtocolVersion": ["1.1"],
                         "subprotocol": "DSP",
                         "subprotocolBody": f"id={asset_id_};dspEndpoint={endpoint_}",
-                        "subprotocolBodyEncoding": "plain"
+                        "subprotocolBodyEncoding": "plain",
+                        "securityAttributes": [
+                            {
+                                "type": "NONE",
+                                "key": "NONE",
+                                "value": "NONE"
+                            }
+                        ]
                     }
                 }
             ]
@@ -320,6 +327,8 @@ if __name__ == "__main__":
     parser.add_argument("--aas3", help="Create AAS assets in version 3.0", action='store_true', required=False)
     parser.add_argument("-d", "--dataplane", type=str, nargs="*", help="EDC provider data plane display URLs",
                         required=False)
+    parser.add_argument("--allowedBPNs", type=str, nargs="*",
+                        help="The allowed BPNs for digital twin registration in the registry.", required=False)
 
     args = parser.parse_args()
     config = vars(args)
@@ -339,6 +348,7 @@ if __name__ == "__main__":
     bpns_list = config.get("bpns")
     is_aas3 = config.get("aas3")
     dataplane_urls = config.get("dataplane")
+    allowedBPNs = config.get("allowedBPNs")
 
     if is_aas3 and dataplane_urls is None:
         raise ArgumentException("Dataplane URLs have to be specified with -d or --dataplane if --aas flag is set!")
@@ -353,6 +363,9 @@ if __name__ == "__main__":
 
     if aas_upload_url is None:
         aas_upload_url = aas_url
+
+    if allowedBPNs is None:
+        allowedBPNs = []
 
     registry_path = "/registry/shell-descriptors"
     if is_aas3:
@@ -450,15 +463,31 @@ if __name__ == "__main__":
                     })
             print(name_at_manufacturer)
 
-            specific_asset_ids_temp.append({
+            manufacturerId = {
                 "key": "manufacturerId",
                 "value": tmp_data["bpnl"]
-            })
+            }
+            if manufacturerId not in specific_asset_ids_temp:
+                specific_asset_ids_temp.append(manufacturerId)
             if is_aas3:
+                keys = [{
+                    "type": "GlobalReference",
+                    "value": "PUBLIC_READABLE"
+                }]
+                for bpn in allowedBPNs:
+                    keys.append({
+                        "type": "GlobalReference",
+                        "value": bpn
+                    })
+                externalSubjectId = {
+                    "type": "ExternalReference",
+                    "keys": keys
+                }
                 for asset in specific_asset_ids_temp:
                     specific_asset_ids.append({
                         "name": asset.get("key"),
-                        "value": asset.get("value")
+                        "value": asset.get("value"),
+                        "externalSubjectId": externalSubjectId
                     })
             else:
                 specific_asset_ids = specific_asset_ids_temp